VariablesModule

Description

Obsahuje funkcie a struktury na pracu s funkciama a premennyma

Documentation

data Expr

Hodnoty premennych

Constructors

Blank

LiteralInt Integer

LiteralDouble Double

LiteralBool Bool

LiteralString String

Var String

Mult Expr Expr

Plus Expr Expr

Minus Expr Expr

Div Expr Expr

GrT Expr Expr

LeT Expr Expr

Equals Expr Expr

NotEquals Expr Expr

GrTE Expr Expr

LeTE Expr Expr

UnMinus Expr

Function String [Expr]

FunctionDef

Fields

fncName ∷ String

fncParams ∷ [(String, Expr, VarTypes)]

retType ∷ VarTypes

body ∷ [Command]

fncVariables ∷ SymbolTable

retValue ∷ Expr

evRetValue ∷ Expr

FunctionDec String [(String, Expr, VarTypes)] VarTypes

data Command

Datova struktura pre prikazy

Constructors

Empty

Assign String Expr

NewVariable String Expr VarTypes

Print Expr

Seq [Command]

Input String

If Expr Command Command

SimpleIf Expr Command

While Expr Command

DeclareFunction Expr

DeclareFncs [Command]

DefineFncs [Command]

DefineFunction Expr

ReturnInScope

CallFunction String [Expr]

Document [Command] [Command]

type SymbolTable = [(String, Expr, VarTypes)]

List (meno,hodnota,typ) premennych

castIntToDouble ∷ Expr → Expr

Pretypuje LiteralInt na LiteralDouble

myDiv ∷ Expr → Expr → Expr

Delenie Expr vyrazov

set ∷ SymbolTable → String → Expr → VarTypes → SymbolTable

Vlozi premennu resp. funkciu do tabulky

get ∷ SymbolTable → String → (Expr, VarTypes)

Vyhlada premennu, nevracia funkcie

getAllFnc ∷ SymbolTable → SymbolTable

Vrati vsetky definicie funkcii

getAllFncDec ∷ SymbolTable → SymbolTable

Vrati vsetky deklaracie funkcii

getFncDef ∷ SymbolTable → String → ([(String, Expr, VarTypes)], VarTypes, [Command], SymbolTable, Expr, Expr)

Vyhlada definiciu funkcie

getFncDefAll ∷ SymbolTable → String → Expr

Vyhlada defniciu funkcie a vrati ju celu ako je ulozena v SymbolTable

isFncDef ∷ SymbolTable → String → Bool

Zisti ci uz je funkcia definovana

getFncDec ∷ SymbolTable → String → Maybe ([(String, Expr, VarTypes)], VarTypes)

Vyhlada deklaraciu funkcie

isFncDec ∷ SymbolTable → String → Bool

Zisti ci je uz funkcia deklarovana

getFncParams ∷ SymbolTable → String → [(String, Expr, VarTypes)]

Vrati vsetky parametre funkcie

getFncRetType ∷ SymbolTable → String → VarTypes

Vrati typ navratovej hodnoty funkcie

getFncBody ∷ SymbolTable → String → [Command]

Vrati telo funkcie (sekvenciu prikazov)

getFncVariables ∷ SymbolTable → String → SymbolTable

Vsetky lokalne premenne funkcie

getFncRetValueExpr ∷ SymbolTable → String → Expr

Vrati hodnotu navratovej hodnoty ako Expr bez vyhodnotenia

getFncRetValue ∷ SymbolTable → SymbolTable → String → Expr

Vrati navratovu hodnotu funkcie a ak je to premenna tak aj vyhodnoti jej hodnotu

getEvaluatedRetValue ∷ SymbolTable → String → Expr

Pouziva sa ak je navratova hodnota premenna, tak vrati hodnotu

getFncDecParams ∷ SymbolTable → String → [(String, Expr, VarTypes)]

Vrati vsetky parametre deklaracia funkcie

getFncDecRetType ∷ SymbolTable → String → VarTypes

Vrati typ navratovej hodnoty deklaracie funkcie

getFncCalls ∷ [Command] → [Command]

Zo vsetkych prikazov vyfiltruje len volania funkcii

fncDecAndDefMatch ∷ SymbolTable → String → Expr → Bool

Skontroluje typy parametrov a navratovu hodnotu definicii a deklaracii funkcii

setFncEvValue ∷ SymbolTable → String → Expr → SymbolTable

Nastavi vyhodnotenu hodnotu do funkcie a skontroluje typ navratovej hodnoty a hodnotu

isDefinitionMissing ∷ SymbolTable → Bool

Skontroluje ci nechyba definicie pre niektoru deklraciu

data VarTypes

Typy premennych pouzivanych pri typovej kontrole

Constructors

BoolVar

IntVar

DoubleVar

StringVar

BlankVar

FunctionVar

FunctionDecVar

FunctionDefVar

VarName

getVarType ∷ SymbolTable → String → Maybe VarTypes

Vrati typ premennej pre meno existujucej premennej

isBlankVar ∷ Expr → Bool

isIntLiteral ∷ Expr → Bool

isDoubleLiteral ∷ Expr → Bool

isBoolLiteral ∷ Expr → Bool

isStringLiteral ∷ Expr → Bool

isBlankLiteral ∷ Expr → Bool

isFunction ∷ Expr → Bool

isVarName ∷ Expr → Bool

isExpr ∷ Expr → Bool

isUnMinusLiteral ∷ Expr → Bool

getUnMinusLiteralType ∷ Expr → Expr

isIntVar ∷ VarTypes → Bool

isDoubleVar ∷ VarTypes → Bool

isBoolVar ∷ VarTypes → Bool

isStringVar ∷ VarTypes → Bool

isFunctionVar ∷ VarTypes → Bool

getValueType ∷ Expr → VarTypes

Vrati typ pre zadanu Expr hodnotu